% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_table.R
\docType{data}
\name{constants_table}
\alias{constants_table}
\title{Fleishman's Power Method Transformation Constants}
\format{
A \code{data.frame} with 5 columns, which are
\describe{
 \item{Skew}{The skewness value}
 \item{Kurtosis}{The standardized kurtosis value}
 \item{b}{Outcome that is based on \code{Skew,Kurtosis}}
 \item{c}{Outcome that is based on \code{Skew,Kurtosis}}
 \item{d}{Outcome that is based on \code{Skew,Kurtosis}}
}
}
\usage{
constants_table
}
\description{
This table includes Fleishman's Power Method
Transformation constants.
}
\examples{

# First 6 rows of the table
data(constants_table)
head(constants_table)

}
\references{
Fleishman AI (1978). A Method for Simulating Non-normal Distributions.
 \emph{Psychometrika, 43, 521-532.} \doi{10.1007/BF02293811}.

Fialkowski, A. C.  (2018). SimMultiCorrData: Simulation of Correlated
Data with Multiple Variable Types.  R package version 0.2.2. Retrieved
from
https://cran.r-project.org/web/packages/SimMultiCorrData/index.html
}
\seealso{
\code{\link[SimMultiCorrData]{find_constants}}
}
\keyword{datasets}
