% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample_shiny.R
\name{draw_sample_shiny}
\alias{draw_sample_shiny}
\title{Draw Samples with a Shiny Applications}
\usage{
draw_sample_shiny()
}
\description{
Performing package functions with user friendly 'shiny' interface.
}
\examples{
\dontrun{
# if(interactive()){
## Run this code for launching the 'shiny' application
#  draw_sample_shiny()
#  }
# }
}
