% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draws}
\alias{draws}
\title{Draw a number from a given distribution}
\usage{
draws(n, dist, ...)
}
\arguments{
\item{n}{Numeric. The number of draws.}

\item{dist}{String. The distribution numbers are going to be drawn from.}

\item{...}{Numeric, parameters of the distribution}
}
\value{
Returns a list of the values drawn from the distribtuion
}
\description{
This function returns a given number of draws from
the specified distribution. The distribution to draw from can be the Normal, Poisson
or Binomial distributions.
}
\examples{
# Sample 100 draws from the normal distribtuion wih mean = 0
# and standard desviation = 1:
draws(100, 'normal', mean = 0, sd = 1)

# Sample 100 draws from the poison distribution with lambda = 1
draws(100, 'poisson', lambda =1)

# Sample 100 draws from the binomial distribution with size (n) = 50
# and probability (p)
draws(100, 'poisson', lambda =1)
}
