% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A \code{data.frame} with 3 columns, which are
\describe{
 \item{ID}{students' id}
 \item{Score_1}{Scores of test 1}
 \item{Score_2}{Scores of test 2}
}
}
\usage{
data(example_data)
}
\description{
The example data set  is made of 500 subjects ids and
total scores from two different tests.
}
\keyword{datasets}
