% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_example.R
\docType{data}
\name{likert_example}
\alias{likert_example}
\title{Likert Example Data}
\format{
A \code{data.frame} with 7 columns, which are
\describe{
 \item{CNTSTUID}{country ID}
 \item{ST160Q01IA}{response of item_1}
 \item{ST160Q02IA}{response of item_2}
 \item{ST160Q03IA}{response of item_3}
 \item{ST160Q04IA}{response of item_4}
 \item{ST160Q05IA}{response of item_5}
 \item{total}{total_score of five items}
}
}
\usage{
data(likert_example)
}
\description{
The example data set  is made of 6669 subjects, 7 variables
}
\examples{
# First 6 rows of the likert_example
data(likert_example)
head(likert_example)
}
\keyword{datasets}
